feat(report): allow slashes in version
Allow the slashes in the version string to not be expaced.

`feat%2Fdashboard` -> `feat/dashboard`

This is better for readability as well plays nice with azure routing.

For now we enforce the project name to have exactly 3 parts, for example  "provider/org/name". We will need to rework that in the future to allow project names like "gitlab.com/org/team/backend/server".