build: use docker actions steps for building image (#578)
Uses github actions steps for building the Docker image, rather than the CLI. The steps are slightly more verbose, but this makes it easier to specify build arguments, tags and labels. For example, release builds are automatically tagged as the tag name, and `:latest`. We could also easily publish to the github container registry with this, if wanted.
Also adds a multi-platform build to run on `linux/arm64`.

- Also adds a cache for the Docker build, to speed up the build process.
- And separates the `npm install` step from the final image, to reduce the image size.